Move lot of things to private & remove Option on QueryPlan
