+++ apollo-router-core/src/response.rs
@@ -1,8 +1,6 @@
 use crate::prelude::graphql::*;
-use crate::query_planner::model::Selection;
 use futures::prelude::*;
 use serde::{Deserialize, Serialize};
-use serde_json::map::Entry;
 use std::pin::Pin;
 use typed_builder::TypedBuilder;
 
@@ -59,34 +57,6 @@ impl Response {
         self.path.is_none()
     }
 
-    pub fn select(
-        &self,
-        path: &Path,
-        selections: &[Selection],
-        schema: &Schema,
-    ) -> Result<Value, FetchError> {
-        let values =
-            self.data
-                .get_at_path(path)
-                .map_err(|err| FetchError::ExecutionPathNotFound {
-                    reason: err.to_string(),
-                })?;
-
-        Ok(Value::Array(
-            values
-                .into_iter()
-                .flat_map(|value| match (value, selections) {
-                    (Value::Object(content), requires) => {
-                        select_object(content, requires, schema).transpose()
-                    }
-                    (_, _) => Some(Err(FetchError::ExecutionInvalidContent {
-                        reason: "not an object".to_string(),
-                    })),
-                })
-                .collect::<Result<Vec<_>, _>>()?,
-        ))
-    }
-
     pub fn insert_data(&mut self, path: &Path, value: Value) -> Result<(), FetchError> {
         let mut nodes =
             self.data
@@ -121,169 +91,10 @@ impl From<Response> for ResponseStream {
     }
 }
 
-fn select_object(
-    content: &Object,
-    selections: &[Selection],
-    schema: &Schema,
-) -> Result<Option<Value>, FetchError> {
-    let mut output = Object::new();
-    for selection in selections {
-        match selection {
-            Selection::Field(field) => {
-                if let Some(value) = select_field(content, field, schema)? {
-                    match output.entry(field.name.to_owned()) {
-                        Entry::Occupied(mut existing) => existing.get_mut().deep_merge(value),
-                        Entry::Vacant(vacant) => {
-                            vacant.insert(value);
-                        }
-                    }
-                }
-            }
-            Selection::InlineFragment(fragment) => {
-                if let Some(Value::Object(value)) =
-                    select_inline_fragment(content, fragment, schema)?
-                {
-                    output.append(&mut value.to_owned())
-                }
-            }
-        };
-    }
-    if output.is_empty() {
-        return Ok(None);
-    }
-    Ok(Some(Value::Object(output)))
-}
-
-fn select_field(
-    content: &Object,
-    field: &Field,
-    schema: &Schema,
-) -> Result<Option<Value>, FetchError> {
-    match (content.get(&field.name), &field.selections) {
-        (Some(Value::Object(child)), Some(selections)) => select_object(child, selections, schema),
-        (Some(value), None) => Ok(Some(value.to_owned())),
-        (None, _) => Err(FetchError::ExecutionFieldNotFound {
-            field: field.name.to_owned(),
-        }),
-        _ => Ok(None),
-    }
-}
-
-fn select_inline_fragment(
-    content: &Object,
-    fragment: &InlineFragment,
-    schema: &Schema,
-) -> Result<Option<Value>, FetchError> {
-    match (&fragment.type_condition, &content.get("__typename")) {
-        (Some(condition), Some(Value::String(typename))) => {
-            if condition == typename || schema.is_subtype(condition, typename) {
-                select_object(content, &fragment.selections, schema)
-            } else {
-                Ok(None)
-            }
-        }
-        (None, _) => select_object(content, &fragment.selections, schema),
-        (_, None) => Err(FetchError::ExecutionFieldNotFound {
-            field: "__typename".to_string(),
-        }),
-        (_, _) => Ok(None),
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use serde_json::json;
-
-    macro_rules! select {
-        ($schema:expr, $content:expr $(,)?) => {{
-            let schema: Schema = $schema.parse().unwrap();
-            let response = Response::builder()
-                .data($content)
-                .build();
-            let stub = json!([
-                {
-                    "kind": "InlineFragment",
-                    "typeCondition": "OtherStuffToIgnore",
-                    "selections": [],
-                },
-                {
-                    "kind": "InlineFragment",
-                    "typeCondition": "User",
-                    "selections": [
-                        {
-                            "kind": "Field",
-                            "name": "__typename",
-                        },
-                        {
-                            "kind": "Field",
-                            "name": "id",
-                        },
-                        {
-                            "kind": "Field",
-                            "name": "job",
-                            "selections": [
-                                {
-                                    "kind": "Field",
-                                    "name": "name",
-                                }
-                            ],
-                        }
-                      ]
-                },
-            ]);
-            let selection: Vec<Selection> = serde_json::from_value(stub).unwrap();
-            response.select(&Path::empty(), &selection, &schema)
-        }};
-    }
-
-    #[test]
-    fn test_selection() {
-        assert_eq!(
-            select!(
-                "",
-                json!({"__typename": "User", "id":2, "name":"Bob", "job":{"name":"astronaut"}}),
-            )
-            .unwrap(),
-            json!([{
-                "__typename": "User",
-                "id": 2,
-                "job": {
-                    "name": "astronaut"
-                }
-            }]),
-        );
-    }
-
-    #[test]
-    fn test_selection_subtype() {
-        assert_eq!(
-            select!(
-                "union User = Author | Reviewer",
-                json!({"__typename": "Author", "id":2, "name":"Bob", "job":{"name":"astronaut"}}),
-            )
-            .unwrap(),
-            json!([{
-                "__typename": "Author",
-                "id": 2,
-                "job": {
-                    "name": "astronaut"
-                }
-            }]),
-        );
-    }
-
-    #[test]
-    fn test_selection_missing_field() {
-        assert!(matches!(
-            select!(
-                "",
-                json!({"__typename": "User", "name":"Bob", "job":{"name":"astronaut"}}),
-            )
-                .unwrap_err(),
-            FetchError::ExecutionFieldNotFound { field } if field == "id"
-        ));
-    }
 
     #[test]
     fn test_insert_data() {
